Refactoring file structure and names in app/javascript
