+++ app/javascript/components/analytics/mapData.js
@@ -1,150 +1,7 @@
-export const totalMonitoreesMap = {
-  columns: {
-    geo: {
-      dataTable: true,
-      label: '',
-      name: 'State',
-      tooltip: false,
-    },
-    primary: {
-      dataTable: true,
-      label: 'Total Monitorees',
-      name: 'Total',
-      prefix: '',
-      suffix: '',
-      tooltip: true,
-    },
-    navigate: {
-      dataTable: false,
-      name: 'URL',
-      tooltip: false,
-    },
-  },
-  legend: {
-    numberOfItems: 4,
-    position: 'side',
-    title: 'Count of Monitorees',
-    type: 'equalnumber',
-    specialClasses: ['Sara Alert Not in Use'],
-    data: [],
-  },
-  data: [{ State: null }],
-  general: {
-    title: 'All Monitorees By Location Over Time',
-    type: 'data',
-    geoType: 'us',
-    headerColor: 'theme-slate',
-    showSidebar: true,
-    showDownloadButton: true,
-    expandDataTable: false,
-  },
-  color: 'orangered',
-};
+/**
+ * Objects used for configuring and populating analytics maps.
+ */
 
-export const symptomaticMonitoreesMap = {
-  columns: {
-    geo: {
-      dataTable: true,
-      label: '',
-      name: 'State',
-      tooltip: false,
-    },
-    primary: {
-      dataTable: true,
-      label: 'Symptomatic Monitorees',
-      name: 'Symptomatic',
-      prefix: '',
-      suffix: '',
-      tooltip: true,
-    },
-    navigate: {
-      dataTable: false,
-      name: 'URL',
-      tooltip: false,
-    },
-  },
-  legend: {
-    numberOfItems: 4,
-    position: 'side',
-    title: 'Count of Symptomatic Monitorees',
-    type: 'equalnumber',
-    specialClasses: ['Sara Alert Not in Use'],
-    data: [],
-  },
-  data: [{ State: null }],
-  general: {
-    title: 'Symptomatic Monitorees By Location Over Time',
-    type: 'data',
-    geoType: 'us',
-    headerColor: 'theme-slate',
-    showSidebar: true,
-    showDownloadButton: true,
-    expandDataTable: false,
-  },
-  color: 'orangered',
-};
-
-export const stateOptions = [
-  { abbrv: 'AL', isoCode: 'US-AL', mapFile: 'al', name: 'Alabama' },
-  { abbrv: 'AK', isoCode: 'US-AK', mapFile: 'ak', name: 'Alaska' },
-  { abbrv: 'AS', isoCode: 'US-AS', mapFile: 'usaTerritories', name: 'American Samoa' },
-  { abbrv: 'AZ', isoCode: 'US-AZ', mapFile: 'az', name: 'Arizona' },
-  { abbrv: 'AR', isoCode: 'US-AR', mapFile: 'ar', name: 'Arkansas' },
-  { abbrv: 'CA', isoCode: 'US-CA', mapFile: 'ca', name: 'California' },
-  { abbrv: 'CO', isoCode: 'US-CO', mapFile: 'co', name: 'Colorado' },
-  { abbrv: 'CT', isoCode: 'US-CT', mapFile: 'ct', name: 'Connecticut' },
-  { abbrv: 'DE', isoCode: 'US-DE', mapFile: 'de', name: 'Delaware' },
-  { abbrv: 'DC', isoCode: 'US-DC', mapFile: 'dc', name: 'District of Columbia' },
-  { abbrv: 'FM', isoCode: 'US-FM', mapFile: 'usaTerritories', name: 'Federated States of Micronesia' },
-  { abbrv: 'FL', isoCode: 'US-FL', mapFile: 'fl', name: 'Florida' },
-  { abbrv: 'GA', isoCode: 'US-GA', mapFile: 'ga', name: 'Georgia' },
-  { abbrv: 'GU', isoCode: 'US-GU', mapFile: 'usaTerritories', name: 'Guam' },
-  { abbrv: 'HI', isoCode: 'US-HI', mapFile: 'hi', name: 'Hawaii' },
-  { abbrv: 'ID', isoCode: 'US-ID', mapFile: 'id', name: 'Idaho' },
-  { abbrv: 'IL', isoCode: 'US-IL', mapFile: 'il', name: 'Illinois' },
-  { abbrv: 'IN', isoCode: 'US-IN', mapFile: 'in', name: 'Indiana' },
-  { abbrv: 'IA', isoCode: 'US-IA', mapFile: 'ia', name: 'Iowa' },
-  { abbrv: 'KS', isoCode: 'US-KS', mapFile: 'ks', name: 'Kansas' },
-  { abbrv: 'KY', isoCode: 'US-KY', mapFile: 'ky', name: 'Kentucky' },
-  { abbrv: 'LA', isoCode: 'US-LA', mapFile: 'la', name: 'Louisiana' },
-  { abbrv: 'ME', isoCode: 'US-ME', mapFile: 'me', name: 'Maine' },
-  { abbrv: 'MH', isoCode: 'MH', mapFile: 'usaTerritories', name: 'Marshall Islands' },
-  { abbrv: 'MD', isoCode: 'US-MD', mapFile: 'md', name: 'Maryland' },
-  { abbrv: 'MA', isoCode: 'US-MA', mapFile: 'ma', name: 'Massachusetts' },
-  { abbrv: 'MI', isoCode: 'US-MI', mapFile: 'mi', name: 'Michigan' },
-  { abbrv: 'MN', isoCode: 'US-MN', mapFile: 'mn', name: 'Minnesota' },
-  { abbrv: 'MS', isoCode: 'US-MS', mapFile: 'ms', name: 'Mississippi' },
-  { abbrv: 'MO', isoCode: 'US-MO', mapFile: 'mo', name: 'Missouri' },
-  { abbrv: 'MT', isoCode: 'US-MT', mapFile: 'mt', name: 'Montana' },
-  { abbrv: 'NE', isoCode: 'US-NE', mapFile: 'ne', name: 'Nebraska' },
-  { abbrv: 'NV', isoCode: 'US-NV', mapFile: 'nv', name: 'Nevada' },
-  { abbrv: 'NH', isoCode: 'US-NH', mapFile: 'nh', name: 'New Hampshire' },
-  { abbrv: 'NJ', isoCode: 'US-NJ', mapFile: 'nj', name: 'New Jersey' },
-  { abbrv: 'NM', isoCode: 'US-NM', mapFile: 'nm', name: 'New Mexico' },
-  { abbrv: 'NY', isoCode: 'US-NY', mapFile: 'ny', name: 'New York' },
-  { abbrv: 'NC', isoCode: 'US-NC', mapFile: 'nc', name: 'North Carolina' },
-  { abbrv: 'ND', isoCode: 'US-ND', mapFile: 'nd', name: 'North Dakota' },
-  { abbrv: 'MP', isoCode: 'US-MP', mapFile: 'usaTerritories', name: 'Northern Mariana Islands' },
-  { abbrv: 'OH', isoCode: 'US-OH', mapFile: 'oh', name: 'Ohio' },
-  { abbrv: 'OK', isoCode: 'US-OK', mapFile: 'ok', name: 'Oklahoma' },
-  { abbrv: 'OR', isoCode: 'US-OR', mapFile: 'or', name: 'Oregon' },
-  { abbrv: 'PW', isoCode: 'PW', mapFile: 'usaTerritories', name: 'Palau' },
-  { abbrv: 'PA', isoCode: 'US-PA', mapFile: 'pa', name: 'Pennsylvania' },
-  { abbrv: 'PR', isoCode: 'US-PR', mapFile: 'usaTerritories', name: 'Puerto Rico' },
-  { abbrv: 'RI', isoCode: 'US-RI', mapFile: 'ri', name: 'Rhode Island' },
-  { abbrv: 'SC', isoCode: 'US-SC', mapFile: 'sc', name: 'South Carolina' },
-  { abbrv: 'SD', isoCode: 'US-SD', mapFile: 'sd', name: 'South Dakota' },
-  { abbrv: 'TN', isoCode: 'US-TN', mapFile: 'tn', name: 'Tennessee' },
-  { abbrv: 'TX', isoCode: 'US-TX', mapFile: 'tx', name: 'Texas' },
-  { abbrv: 'UT', isoCode: 'US-UT', mapFile: 'ut', name: 'Utah' },
-  { abbrv: 'VT', isoCode: 'US-VT', mapFile: 'vt', name: 'Vermont' },
-  { abbrv: 'VI', isoCode: 'US-VI', mapFile: 'usaTerritories', name: 'Virgin Islands' },
-  { abbrv: 'VA', isoCode: 'US-VA', mapFile: 'va', name: 'Virginia' },
-  { abbrv: 'WA', isoCode: 'US-WA', mapFile: 'wa', name: 'Washington' },
-  { abbrv: 'WV', isoCode: 'US-WV', mapFile: 'wv', name: 'West Virginia' },
-  { abbrv: 'WI', isoCode: 'US-WI', mapFile: 'wi', name: 'Wisconsin' },
-  { abbrv: 'WY', isoCode: 'US-WY', mapFile: 'wy', name: 'Wyoming' },
-];
 
 // Amcharts gets finicky with its internal map projections using certain iso codes, hence not all are `US-XX`
 export const insularAreas = [
@@ -156,258 +13,6 @@ export const insularAreas = [
   { abbrv: 'PW', isoCode: 'PW', name: 'Palau', mapFile: 'usaTerritories' },
   { abbrv: 'PR', isoCode: 'US-PR', name: 'Puerto Rico', mapFile: 'usaTerritories' },
   { abbrv: 'VI', isoCode: 'US-VI', name: 'Virgin Islands', mapFile: 'usaTerritories' },
-];
-
-export const countryOptions = [
-  'Afghanistan',
-  'Albania',
-  'Algeria',
-  'American Samoa',
-  'Andorra',
-  'Angola',
-  'Anguilla',
-  'Antarctica',
-  'Antigua and Barbuda',
-  'Argentina',
-  'Armenia',
-  'Aruba',
-  'Australia',
-  'Austria',
-  'Azerbaijan',
-  'Bahamas (the)',
-  'Bahrain',
-  'Bangladesh',
-  'Barbados',
-  'Belarus',
-  'Belgium',
-  'Belize',
-  'Benin',
-  'Bermuda',
-  'Bhutan',
-  'Bolivia (Plurinational State of)',
-  'Bonaire, Sint Eustatius and Saba',
-  'Bosnia and Herzegovina',
-  'Botswana',
-  'Bouvet Island',
-  'Brazil',
-  'British Indian Ocean Territory (the)',
-  'Brunei Darussalam',
-  'Bulgaria',
-  'Burkina Faso',
-  'Burundi',
-  'Cabo Verde',
-  'Cambodia',
-  'Cameroon',
-  'Canada',
-  'Cayman Islands (the)',
-  'Central African Republic (the)',
-  'Chad',
-  'Chile',
-  'China',
-  'Christmas Island',
-  'Cocos (Keeling) Islands (the)',
-  'Colombia',
-  'Comoros (the)',
-  'Congo (the Democratic Republic of the)',
-  'Congo (the)',
-  'Cook Islands (the)',
-  'Costa Rica',
-  'Croatia',
-  'Cuba',
-  'Curaçao',
-  'Cyprus',
-  'Czechia',
-  "Côte d'Ivoire",
-  'Denmark',
-  'Djibouti',
-  'Dominica',
-  'Dominican Republic (the)',
-  'Ecuador',
-  'Egypt',
-  'El Salvador',
-  'Equatorial Guinea',
-  'Eritrea',
-  'Estonia',
-  'Eswatini',
-  'Ethiopia',
-  'Falkland Islands (the) [Malvinas]',
-  'Faroe Islands (the)',
-  'Fiji',
-  'Finland',
-  'France',
-  'French Guiana',
-  'French Polynesia',
-  'French Southern Territories (the)',
-  'Gabon',
-  'Gambia (the)',
-  'Georgia',
-  'Germany',
-  'Ghana',
-  'Gibraltar',
-  'Greece',
-  'Greenland',
-  'Grenada',
-  'Guadeloupe',
-  'Guam',
-  'Guatemala',
-  'Guernsey',
-  'Guinea',
-  'Guinea-Bissau',
-  'Guyana',
-  'Haiti',
-  'Heard Island and McDonald Islands',
-  'Holy See (the)',
-  'Honduras',
-  'Hong Kong',
-  'Hungary',
-  'Iceland',
-  'India',
-  'Indonesia',
-  'Iran (Islamic Republic of)',
-  'Iraq',
-  'Ireland',
-  'Isle of Man',
-  'Israel',
-  'Italy',
-  'Jamaica',
-  'Japan',
-  'Jersey',
-  'Jordan',
-  'Kazakhstan',
-  'Kenya',
-  'Kiribati',
-  "Korea (the Democratic People's Republic of)",
-  'Korea (the Republic of)',
-  'Kuwait',
-  'Kyrgyzstan',
-  "Lao People's Democratic Republic (the)",
-  'Latvia',
-  'Lebanon',
-  'Lesotho',
-  'Liberia',
-  'Libya',
-  'Liechtenstein',
-  'Lithuania',
-  'Luxembourg',
-  'Macao',
-  'Madagascar',
-  'Malawi',
-  'Malaysia',
-  'Maldives',
-  'Mali',
-  'Malta',
-  'Marshall Islands (the)',
-  'Martinique',
-  'Mauritania',
-  'Mauritius',
-  'Mayotte',
-  'Mexico',
-  'Micronesia (Federated States of)',
-  'Moldova (the Republic of)',
-  'Monaco',
-  'Mongolia',
-  'Montenegro',
-  'Montserrat',
-  'Morocco',
-  'Mozambique',
-  'Myanmar',
-  'Namibia',
-  'Nauru',
-  'Nepal',
-  'Netherlands (the)',
-  'New Caledonia',
-  'New Zealand',
-  'Nicaragua',
-  'Niger (the)',
-  'Nigeria',
-  'Niue',
-  'Norfolk Island',
-  'Northern Mariana Islands (the)',
-  'Norway',
-  'Oman',
-  'Pakistan',
-  'Palau',
-  'Palestine, State of',
-  'Panama',
-  'Papua New Guinea',
-  'Paraguay',
-  'Peru',
-  'Philippines (the)',
-  'Pitcairn',
-  'Poland',
-  'Portugal',
-  'Puerto Rico',
-  'Qatar',
-  'Republic of North Macedonia',
-  'Romania',
-  'Russian Federation (the)',
-  'Rwanda',
-  'Réunion',
-  'Saint Barthélemy',
-  'Saint Helena, Ascension and Tristan da Cunha',
-  'Saint Kitts and Nevis',
-  'Saint Lucia',
-  'Saint Martin (French part)',
-  'Saint Pierre and Miquelon',
-  'Saint Vincent and the Grenadines',
-  'Samoa',
-  'San Marino',
-  'Sao Tome and Principe',
-  'Saudi Arabia',
-  'Senegal',
-  'Serbia',
-  'Seychelles',
-  'Sierra Leone',
-  'Singapore',
-  'Sint Maarten (Dutch part)',
-  'Slovakia',
-  'Slovenia',
-  'Solomon Islands',
-  'Somalia',
-  'South Africa',
-  'South Georgia and the South Sandwich Islands',
-  'South Sudan',
-  'Spain',
-  'Sri Lanka',
-  'Sudan (the)',
-  'Suriname',
-  'Svalbard and Jan Mayen',
-  'Sweden',
-  'Switzerland',
-  'Syrian Arab Republic',
-  'Taiwan (Province of China)',
-  'Tajikistan',
-  'Tanzania, United Republic of',
-  'Thailand',
-  'Timor-Leste',
-  'Togo',
-  'Tokelau',
-  'Tonga',
-  'Trinidad and Tobago',
-  'Tunisia',
-  'Turkey',
-  'Turkmenistan',
-  'Turks and Caicos Islands (the)',
-  'Tuvalu',
-  'Uganda',
-  'Ukraine',
-  'United Arab Emirates (the)',
-  'United Kingdom of Great Britain and Northern Ireland (the)',
-  'United States Minor Outlying Islands (the)',
-  'United States of America (the)',
-  'Uruguay',
-  'Uzbekistan',
-  'Vanuatu',
-  'Venezuela (Bolivarian Republic of)',
-  'Viet Nam',
-  'Virgin Islands (British)',
-  'Virgin Islands (U.S.)',
-  'Wallis and Futuna',
-  'Western Sahara',
-  'Yemen',
-  'Zambia',
-  'Zimbabwe',
-  'Åland Islands',
 ];
 
 export const USAMap = {
