+++ app/javascript/components/enrollment/steps/Address.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import { Card, Button, Tabs, Tab, Form, Col } from 'react-bootstrap';
-import { stateOptions, countryOptions } from '../../data';
+import { countryOptions } from '../../../data/countryOptions';
+import { stateOptions } from '../../../data/stateOptions'
 import { PropTypes } from 'prop-types';
 import * as yup from 'yup';
