+++ app/javascript/components/patient/Patient.js
